+++ back-end/controller/index.js
@@ -1,3 +1,4 @@
 import { authorController } from "./author.js";
+import { bookController } from "./book.js";
 
-export { authorController };
+export { authorController, bookController };