+++ back-end/database/connection.js
@@ -1,12 +1,12 @@
 import mongoose from "mongoose";
 
- // DB config
+// DB config
 const dbConnection = async () => {
   try {
     await mongoose.connect(process.env.CONNECTION_URL, {
       useNewUrlParser: true,
       useUnifiedTopology: true,
-      dbName:"lms_project"
+      dbName: "lms_project",
     });
     console.log("Mongodb connected!");
   } catch (error) {