+++ back-end/dto/BookDTO.js
@@ -1,0 +1,12 @@
+class BookDto {
+  name;
+  isbn;
+  authorId;
+
+  constructor(book) {
+    this.name = book.name;
+    this.isbn = book.isbn;
+    this.authorId = book.authorId;
+  }
+}
+export default BookDto;