+++ back-end/Changelog.md
@@ -4,6 +4,8 @@
 ## [1.0.0] - 2022-10-26
 ### Added
 
+____________Author______________
+
 - Config db configurations and api security
 - Create author,handle validation and exception handling
 - Create get all authors function and implement dto
@@ -11,3 +13,6 @@
 - make change the find author by id function(exeption handling)
 - write code comments for the all of author things
 - create find author by id and update function
+
+____________Book______________________
+- Create add new book and validate